python_utils/pyproject.toml
[tool.poetry]
name = "indexify-python-utils"
version = "0.1.0"
description = "Common utilities for Python packages"
authors = ["Tensorlake Inc. <support@tensorlake.ai>"]
license = "Apache 2.0"
readme = "README.md"
homepage = "https://github.com/tensorlakeai/indexify"
repository = "https://github.com/tensorlakeai/indexify"
packages = [
    { include = "python_utils", from="src" },
]

[tool.poetry.dependencies]
python = "^3.9"
pyyaml = "^6"
structlog = "^24.4.0"
httpx = { version = "0.27.2", extras = ["http2"] }

[tool.poetry.dev-dependencies]
black = "^24.10.0"
pylint = "^2.4.0"

[build-system]
requires = ["poetry>=1.2"]
build-backend = "poetry.core.masonry.api"
